requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv